catalogue_graph/pyproject.toml
[project]
name = "catalogue-graph"
version = "0.1.0"
description = "Catalogue graph processing and indexing"
readme = "README.md"
# This project requires Python 3.12 instead of 3.13
# Wheels for pyarrow and pyiceberg are not available for 3.13 yet (as of August 2025)
requires-python = ">=3.12,<3.13"
dependencies = [
    "boto3",
    "requests",
    "backoff",
    "smart-open",
    "polars",
    "elasticsearch>=8.11,<8.13",
    "tabulate",
    "types-tabulate",
    "types-requests",
    "python-dateutil",
    "types-python-dateutil",
    "pyarrow~=20.0",
    "lxml>=6.0.0",
    # Pin pydantic below 2.12.0 due to pyiceberg TableMetadata validation incompatibility
    # (AttributeError: ValidationInfo has no attribute current_schema_id with pydantic 2.12.0)
    "pydantic>=2.11.7,<2.12.0",
    "pyiceberg[sql-sqlite,pyiceberg-core]>=0.9.1",
    "pymarc>=5.3.1",
    "smart-open>=7.3.0.post1",
    "tqdm>=4.67.1",
    "oai-pmh-client",
    "structlog>=24.1.0",
    "pip-system-certs>=5.3",
]

[tool.uv]
package = true

[tool.uv.sources]
oai-pmh-client = { git = "https://github.com/wellcomecollection/oai-pmh.git", tag = "v1.0.1" }


[dependency-groups]
dev = [
    "boto3-stubs",
    "coverage[toml]",
    "mypy>=1.16.1",
    "pytest",
    "pytest-cov",
    "freezegun",
    "ruff",
    "polars>=1.32.3",
    "ipykernel>=6.30.0",
    "pandas>=2.3.1",
    "lxml-stubs>=0.5.1",
    "pyarrow-stubs",
    "pytest>=8.4.0",
    "ipykernel>=6.30.0",
    "pytest-bdd>=8.1.0",
    "hypothesis>=6.100.0,<7.0.0",
    "jinja2>=3.1.6",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests", "integration"]
addopts = "--cov src --cov-report term --cov-report xml:coverage.xml -m 'not integration'"
filterwarnings = "ignore::DeprecationWarning"
markers = [
    "integration: integration tests (using production graph database)",
]

[tool.ruff]
line-length = 88
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I"
]
ignore = ["E501", "B019"]

[tool.mypy]
strict_optional = true
strict_equality = true
disallow_untyped_calls = true
disallow_untyped_defs = true

warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
mypy_path = "src"
exclude = ["build/.*"]

# Ignore untyped imports for external libraries without stubs
[[tool.mypy.overrides]]
module = ["pyarrow.*", "boto3.*", "smart_open.*", "pymarc.*", "oai_pmh_client.*"]
ignore_missing_imports = true
